requirements.txt
Flask
python-dotenv
pandas
xlrd
openpyxl
Faker